fix type string | string[]
