+++ src/events/schema.js
@@ -16,3 +16,5 @@ export const dragSelectEventSchema = selectDefaultSchema.extend({
   draggable: z.boolean().default(false),
   onDragSelect: z.function().optional(),
 });
+
+export const focusFitIdsSchema = z.union([z.string(), z.array(z.string())]).nullish();
